[client] | refactor(TagVideos): styles.tsx -> style.tsx 로 변경
